Our Last Group Project
Hiya guys, this is our last project, unfortunately. Basically, I wish we would be changing our strategies in our coding procedures. I suggest that we read the lines of our task carefully, and design the structures at first on a piece of paper, before starting to write our codes. The reason I suggest that is because we had devoted lots of time in debugging, which is quite inefficient, showing that we are almost clueless about the basic structures. If we design the structures at first, we wouldn't have been that clueless, like a daft prick trying aimlessly different kinds of possibilities of achieving our objectives. The other suggestion is that we comment everything as soon as we finish a small chunk of codes. I've tried this in my last individual project, and it really helped me explain things in detail. I mean, if we're trying to comment at the very end of our programming procedure, we wouldn't have much time to check whether our comments explain our objectives clearly, since everything's in a rush. Anyway, I really appreciate the effort you've made, and I hope that Simon would be satisfied with our work.